requirements.txt
# Core dependencies
streamlit>=1.28.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI/ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
huggingface-hub>=0.19.0

# OCR
pytesseract>=0.3.10

# LLM for compliance
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Image processing
rembg>=2.0.50
scikit-image>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

